1.Changed Driver.toString 2.Changed Manufacturer.toString 3.Removed Main class 4.Created controller for exception 5.Created enum for home_page and exception_page 6.Renamed jsp files 7.Added validation id in deleteControllers

+++ src/main/java/mate/controller/DeleteCarController.java
@@ -7,17 +7,20 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import mate.lib.Injector;
 import mate.service.CarService;
+import mate.service.Validator;
 
 public class DeleteCarController extends HttpServlet {
-    private static final String MAIN_PAGE = "/index";
     private static final Injector injector = Injector.getInstance("mate");
     private final CarService carService = (CarService) injector.getInstance(CarService.class);
 
     @Override
     public void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        Long id = Long.valueOf(req.getParameter("id"));
+        Long id = Validator.getId(req, resp);
+        if (id == null) {
+            return;
+        }
         carService.delete(id);
-        resp.sendRedirect(MAIN_PAGE);
+        resp.sendRedirect(PageAddress.HOME_PAGE.getPath());
     }
 }
